refactor: move all test files to test/ directory
- Moved 6 spec files from src/ co-location to test/ folder
- Updated all import paths and jest.mock paths accordingly
- Matches original repo convention
